requirements.txt
customtkinter==5.0.3
watchdog==2.2.1
Pillow==9.4.0
tkinter
logging
socket
struct
random
subprocess
